requirements.txt
requests
ollama
pyyaml